format syscalls and make file names arch specifc
