+++ scripts/src.build.sh
@@ -22,33 +22,30 @@ fi
 
 LD_FLAGS="-X github.com/docker-slim/docker-slim/pkg/version.appVersionTag=${TAG} -X github.com/docker-slim/docker-slim/pkg/version.appVersionRev=${REVISION} -X github.com/docker-slim/docker-slim/pkg/version.appVersionTime=${BUILD_TIME}"
 
-# gox -osarch="linux/amd64" -ldflags "${LD_FLAGS}" -output "${BDIR_GOPATH}/bin/linux/docker-slim" 
-# gox -osarch="darwin/amd64" -ldflags "${LD_FLAGS}" -output "${BDIR_GOPATH}/bin/mac/docker-slim"
+gox -verbose -osarch="linux/amd64" -ldflags "${LD_FLAGS}" -output "${BDIR_GOPATH}/bin/linux/docker-slim" 
+gox -verbose -osarch="darwin/amd64" -ldflags "${LD_FLAGS}" -output "${BDIR_GOPATH}/bin/mac/docker-slim"
 gox -osarch="linux/arm" -output "$BDIR_GOPATH/bin/linux_arm/docker-slim"
 popd
 pushd ${BDIR_GOPATH}/cmd/docker-slim-sensor
-# gox -osarch="linux/amd64" -output="${BDIR_GOPATH}/bin/linux/docker-slim-sensor"
+gox -osarch="linux/amd64" -output="${BDIR_GOPATH}/bin/linux/docker-slim-sensor"
 gox -osarch="linux/arm" -output "$BDIR_GOPATH/bin/linux_arm/docker-slim-sensor"
 popd
-# rm -rfv ${BDIR_GOPATH}/dist_mac
-# mkdir ${BDIR_GOPATH}/dist_mac
-# cp ${BDIR_GOPATH}/bin/mac/docker-slim ${BDIR_GOPATH}/dist_mac/docker-slim
-# cp ${BDIR_GOPATH}/bin/linux/docker-slim-sensor ${BDIR_GOPATH}/dist_mac/docker-slim-sensor
-# pushd ${BDIR_GOPATH}
-# zip -r dist_mac.zip dist_mac -x "*.DS_Store"
-# popd
-# rm -rfv ${BDIR_GOPATH}/dist_linux
-# mkdir ${BDIR_GOPATH}/dist_linux
-# cp ${BDIR_GOPATH}/bin/linux/docker-slim ${BDIR_GOPATH}/dist_linux/docker-slim
-# cp ${BDIR_GOPATH}/bin/linux/docker-slim-sensor ${BDIR_GOPATH}/dist_linux/docker-slim-sensor
+rm -rfv ${BDIR_GOPATH}/dist_mac
+mkdir ${BDIR_GOPATH}/dist_mac
+cp ${BDIR_GOPATH}/bin/mac/docker-slim ${BDIR_GOPATH}/dist_mac/docker-slim
+cp ${BDIR_GOPATH}/bin/linux/docker-slim-sensor ${BDIR_GOPATH}/dist_mac/docker-slim-sensor
+pushd ${BDIR_GOPATH}
+zip -r dist_mac.zip dist_mac -x "*.DS_Store"
+popd
+rm -rfv ${BDIR_GOPATH}/dist_linux
+mkdir ${BDIR_GOPATH}/dist_linux
+cp ${BDIR_GOPATH}/bin/linux/docker-slim ${BDIR_GOPATH}/dist_linux/docker-slim
+cp ${BDIR_GOPATH}/bin/linux/docker-slim-sensor ${BDIR_GOPATH}/dist_linux/docker-slim-sensor
 rm -rfv $BDIR_GOPATH/dist_linux_arm
 mkdir $BDIR_GOPATH/dist_linux_arm
 cp $BDIR_GOPATH/bin/linux_arm/docker-slim $BDIR_GOPATH/dist_linux_arm/docker-slim
 cp $BDIR_GOPATH/bin/linux_arm/docker-slim-sensor $BDIR_GOPATH/dist_linux_arm/docker-slim-sensor
-# pushd ${BDIR_GOPATH}
-# tar -czvf dist_linux.tar.gz dist_linux
-# popd
 pushd ${BDIR_GOPATH}
-tar -czvf dist_linux_arm.tar.gz dist_linux_arm
+tar -czvf dist_linux.tar.gz dist_linux
 popd
 rm -rfv ${BDIR_GOPATH}/bin
